Cleanup unused supporters view/serializer
