+++ apps/accounts/tests.py
@@ -58,7 +58,6 @@ class MemberApiIntegrationTest(UserTestsMixin, TestCase):
         """
 
         user_detail_url = "{0}{1}{2}".format(self.member_api_base_url, 'users/', self.user.id)
-        print user_detail_url
         response = self.client.get(user_detail_url)
         self.assertEqual(response.status_code, status.HTTP_200_OK, response.data)
         self.assertEqual(response.data['username'], self.user.username)
@@ -71,8 +70,6 @@ class MemberApiIntegrationTest(UserTestsMixin, TestCase):
 
         # Test unauthenticated user doesn't return 500 error.
         authenticated_member_url = "{0}{1}".format(self.member_api_base_url, 'current')
-        print authenticated_member_url
-
         response = self.client.get(authenticated_member_url)
         self.assertEqual(response.status_code, status.HTTP_200_OK, response.data)
         self.assertEqual(response.data['username'], '')
