+++ apps/projects/urlsapi.py
@@ -1,4 +1,4 @@
-from apps.projects.views import ProjectDonationList, ProjectSupporterList
+from apps.projects.views import ProjectDonationList
 from django.conf.urls import patterns, url, include
 from surlex.dj import surl
 from .views import (ProjectDetail, ProjectList, ProjectWallPostList, ProjectWallPostDetail, ProjectMediaWallPostList,
@@ -20,11 +20,7 @@ urlpatterns = patterns('',
     url(r'^wallposts/media/photos/$', ProjectMediaWallPostPhotoList.as_view(), name='project-mediawallpost-photo-list'),
     surl(r'^wallposts/media/photos/<pk:#>$', ProjectMediaWallPostPhotoDetail.as_view(), name='project-mediawallpost-photo-list'),
 
-    # Project support
-    surl(r'^<slug:s>/donations/$', ProjectDonationList.as_view(), name='project-donation-list'),
-    surl(r'^<slug:s>/supporters/$', ProjectSupporterList.as_view(), name='project-support-list'),
-
+    # Project supporters
     url(r'^donations/$', ProjectDonationList.as_view(), name='project-donation-list'),
-    url(r'^supporters/$', ProjectSupporterList.as_view(), name='project-support-list'),
 
 )
