+++ static/global/js/app/orders.js
@@ -22,17 +22,16 @@ App.Donation = DS.Model.extend({
     order: DS.belongsTo('App.Order')
 });
 
-App.DonationPreview = DS.Model.extend({
-    url: 'fund/previewdonations',
+App.DonationPreview =  DS.Model.extend({
+    url: 'projects/donations',
+    project: DS.belongsTo('App.ProjectPreview'),
+    member: DS.belongsTo('App.MemberPreview'),
 
-    project: DS.belongsTo('App.Project'),
-    user: DS.belongsTo('App.Member'),
-
-    status: DS.attr('string'),
-    type: DS.attr('string'),
+    date_donated: DS.attr('date'),
+    time_since: function(){
+        return Globalize.format(this.get('date_donated'), 'X');
+    }.property('date_donated')
 });
-
-
 
 App.Voucher =  DS.Model.extend({
     url: 'fund/vouchers',
